requirements.txt
flask==3.1.0
flask-cors
flask-session
litellm
pdfminer.six
python-docx
pyyaml
weasyprint
jinja2
scikit-learn
nltk
mcp
daytona-sdk
gunicorn
langgraph
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
langchain-chroma
chromadb
unstructured
beautifulsoup4
validators
pypdf
